omit flag when reading text files on two lines
when omitting the flag with '⎕nget', or when using the '0' flag, the
text will be read as a character vector of all the text rather than an
array of character vectors representing each line.  this means that
extracting the text to pass to the 'ct' or 'ngrams' dfns is now simpler.

previously, when classifying the language of an entire file, the lines
were also joined, which meant that some words weren't tokenized
correctly in the 'ngrams' dfn.